docs(benchmark/bootstrap): expand README.md (#1421)
Currently, anyone wanting to run `aws-launch.sh` will need to peek into
the implementation of the script to understand what it does. For first
time users (like myself), this introduces an extremely large learning
curve.

This PR removes the need to read the implementation of `aws-launch.sh`
by expanding the README.md to cover how to use the script and how to
view the reexecution benchmark results. Using this new PR, I was able to
get an EC2 instance up and running.