[bootstrapmonitor] Switch from `latest` to `master` tag
A recent PR (#4469) proposes to change the `latest` tag to refer to
the image of the most recent release intead of the latest merge to
master. This change requires updating the bootstrap monitor to use the
`master` tag (already updated on every merge to master) to ensure
bootstrap testing continues to target the most recently published image.