refactor(rootstore): replace RootStoreError with boxed error (#1446)
As mentioned in
https://github.com/ava-labs/firewood/pull/1395#discussion_r2504821379,
having to create an instance of `RootStoreError` everytime we want to
return an error is cumbersome. This PR reduces the amount of duplicate
code by removing the `RootStoreError` type as the `method` field isn't
really useful (we can figure out which method returned the error during
debugging) and the `source` field doesn't need to be wrapped.