feat(ffi): add keepalive struct to own waitgroup logic (#1437)
Expanding on Arran and Austin's work, I added the
`databaseKeepAliveHandle` to manage the WaitGroup logic for keeping the
database alive while various objects needed.

In the future, the `RangeProof` object will also use this keep-alive
handle, but only during certain stages of its lifecycle. Because of
this, the disown method on the keep-alive handle is capable of being
called on the nil receiver without causing a panic.