docs(ffi): Update godoc (#1466)
This doesn't totally complete #1113 since I didn't touch the Proof or
Iterator files. However, this does clarify a lot of the use cases, and
what is valid for concurrent operations.

The best way to review this is open the godocs and reading through it,
commenting on what doesn't seem clear. To reproduce the godocs, run:
```
cd ./ffi
godoc -http=:<PORT_NUM>
```

Please leave nit-picky comments! It would be great if any reader of the
docs would know how to use the code.

Closes #1348, #1413