refactor(firewood/db): unify reopen methods (#1481)
IMO, the `reopen()` method of `TestDb` is a misnomer - while this method
does not truncate the existing Firewood file, it uses the default
configuration for the new `TestDb` instance, ignoring the prior
instance's configuration. As a consumer, I would expect `reopen()` to
maintain the existing configuration.

This PR does the following:
- Unifies `reopen()` and `reopen_with_config()` into a single method
- Stores the `dbconfig` in `testDb` to allow the database configuration
to persist across reopens
- Adds documentation to both `reopen()` and `replace()`
- Removes `into_root_store()` test helpers since `RootStore` reopens are
now handled via the database configuration

This PR is a precursor to #1481.